[432]: Registration

+++ apps/shared/scripts.py
@@ -1598,7 +1598,7 @@ def run_inventory_report():
     return True
 
 
-def report_rerun(company_id, start_month, run_fix_data=False):
+def report_rerun(company_id, start_month, run_fix_data=False, run_update_warehouse=False, has_lot=False):
     ReportInventorySub.objects.all().delete()
     ReportInventoryProductWarehouse.objects.all().delete()
     ReportInventory.objects.all().delete()
@@ -1653,29 +1653,30 @@ def report_rerun(company_id, start_month, run_fix_data=False):
             ending_balance_cost=float(8000000),
             for_balance=True
         )
-        ProductWareHouse.objects.filter(id='5da687717c6049a1905a6193e3f41c51').update(
-            stock_amount=5,
-            receipt_amount=5
-        )
-        ProductWareHouse.objects.filter(id='b7b6a84e-f38e-492e-bf9a-7b03cbe02b52').update(
-            stock_amount=14,
-            receipt_amount=26
-        )
+        if run_update_warehouse:
+            ProductWareHouse.objects.filter(id='5da687717c6049a1905a6193e3f41c51').update(
+                stock_amount=5,
+                receipt_amount=5
+            )
+            ProductWareHouse.objects.filter(id='b7b6a84e-f38e-492e-bf9a-7b03cbe02b52').update(
+                stock_amount=14,
+                receipt_amount=26
+            )
 
-        for sn in ProductWareHouseSerial.objects.filter(id__in=[
-            '78275273-0b24-40ee-8f53-c353f640fb12',
-            '8c322c95-116d-4ee4-893c-bb337a0656e1'
-        ]):
-            sn.product_warehouse_id = 'b7b6a84e-f38e-492e-bf9a-7b03cbe02b52'
-            sn.save(update_fields=['product_warehouse_id'])
-
-        prd_wh = ProductWareHouse.objects.filter(id='025e7896fd874c80befda6f18ed148b1').first()
-        if prd_wh:
-            for sn in ProductWareHouseSerial.objects.filter(product_warehouse=prd_wh):
-                sn.product_warehouse_id = '5da687717c6049a1905a6193e3f41c51'
+            for sn in ProductWareHouseSerial.objects.filter(id__in=[
+                '78275273-0b24-40ee-8f53-c353f640fb12',
+                '8c322c95-116d-4ee4-893c-bb337a0656e1'
+            ]):
+                sn.product_warehouse_id = 'b7b6a84e-f38e-492e-bf9a-7b03cbe02b52'
                 sn.save(update_fields=['product_warehouse_id'])
-            prd_wh.delete()
-        ProductWareHouse.objects.filter(id='8d34b13a63c8464daeb4afb53348003a').delete()
+
+            prd_wh = ProductWareHouse.objects.filter(id='025e7896fd874c80befda6f18ed148b1').first()
+            if prd_wh:
+                for sn in ProductWareHouseSerial.objects.filter(product_warehouse=prd_wh):
+                    sn.product_warehouse_id = '5da687717c6049a1905a6193e3f41c51'
+                    sn.save(update_fields=['product_warehouse_id'])
+                prd_wh.delete()
+            ProductWareHouse.objects.filter(id='8d34b13a63c8464daeb4afb53348003a').delete()
         # đầu kỳ Vision
         ReportInventoryProductWarehouse.objects.create(
             tenant=company.tenant,
@@ -1684,6 +1685,7 @@ def report_rerun(company_id, start_month, run_fix_data=False):
             employee_inherit_id='e37c69ca-c5a0-45ff-9c55-dc0bc37b476e',
             product_id='52e45d5b-d91e-4c04-8b2d-7a09ee4820dd',
             warehouse_id='bbac9cfc-df1b-4ed4-97c9-a57ac5c94f89',
+            lot_mapped_id='12de4425e1e341a0bd286e44d78ec260',
             period_mapped_id='5c7423ae29824f338dc5fd2c41b694bf',
             sub_period_order=3,
             sub_period_id='5e1c3cccb4c8439d9b3936a69b72b42a',
@@ -1695,23 +1697,25 @@ def report_rerun(company_id, start_month, run_fix_data=False):
             ending_balance_cost=float(40000000),
             for_balance=True
         )
-        ProductWareHouseLot.objects.filter(id='9b310c99-f2bb-4f78-8f9b-8ec4abdf9023').update(
-            product_warehouse_id='5c296dbf-1885-47b1-aada-b318724d6859'
-        )
-        ProductWareHouseLot.objects.filter(id='12de4425-e1e3-41a0-bd28-6e44d78ec260').update(
-            quantity_import=2
-        )
-        ProductWareHouseLot.objects.filter(id='153cb1a1-7004-4fbd-a08a-61fc9b172a1d').update(
-            quantity_import=3
-        )
-        ProductWareHouse.objects.filter(id='5c296dbf-1885-47b1-aada-b318724d6859').update(
-            sold_amount=12,
-            stock_amount=3
-        )
-        ProductWareHouse.objects.filter(id='52614125-cebd-405c-a59a-a0c4890abfed').update(
-            stock_amount=3,
-            receipt_amount=3
-        )
+        if run_update_warehouse:
+            ProductWareHouseLot.objects.filter(id='9b310c99-f2bb-4f78-8f9b-8ec4abdf9023').update(
+                product_warehouse_id='5c296dbf-1885-47b1-aada-b318724d6859',
+                quantity_import=0
+            )
+            ProductWareHouseLot.objects.filter(id='12de4425-e1e3-41a0-bd28-6e44d78ec260').update(
+                quantity_import=3
+            )
+            ProductWareHouseLot.objects.filter(id='153cb1a1-7004-4fbd-a08a-61fc9b172a1d').update(
+                quantity_import=3
+            )
+            ProductWareHouse.objects.filter(id='5c296dbf-1885-47b1-aada-b318724d6859').update(
+                sold_amount=12,
+                stock_amount=3
+            )
+            ProductWareHouse.objects.filter(id='52614125-cebd-405c-a59a-a0c4890abfed').update(
+                stock_amount=3,
+                receipt_amount=3
+            )
         # đầu kỳ HP
         ReportInventoryProductWarehouse.objects.create(
             tenant=company.tenant,
@@ -1731,69 +1735,111 @@ def report_rerun(company_id, start_month, run_fix_data=False):
             ending_balance_cost=float(138000000),
             for_balance=True
         )
-        prd_wh = ProductWareHouse.objects.filter(
-            product_id="e12e4dd2-fb4e-479d-ae8a-c902f3dbc896"
-        ).first()
-        if prd_wh:
-            prd_wh.stock_amount = 5
-            prd_wh.receipt_amount = 28
-            prd_wh.save(update_fields=['stock_amount', 'receipt_amount'])
-            if ProductWareHouseSerial.objects.filter(vendor_serial_number='hp-dk').count() == 0:
-                new_sn = [
-                    ProductWareHouseSerial(
-                        tenant=company.tenant,
-                        company=company,
-                        vendor_serial_number='hp-dk',
-                        serial_number='hp-dk-01',
-                        product_warehouse=prd_wh
-                    ),
-                    ProductWareHouseSerial(
-                        tenant=company.tenant,
-                        company=company,
-                        vendor_serial_number='hp-dk',
-                        serial_number='hp-dk-02',
-                        product_warehouse=prd_wh
-                    )
-                ]
-                ProductWareHouseSerial.objects.bulk_create(new_sn)
-                prd_wh.product.stock_amount = 5
-                prd_wh.product.available_amount = -24
-                prd_wh.product.save(update_fields=['stock_amount', 'available_amount'])
+        if run_update_warehouse:
+            prd_wh = ProductWareHouse.objects.filter(
+                product_id="e12e4dd2-fb4e-479d-ae8a-c902f3dbc896"
+            ).first()
+            if prd_wh:
+                prd_wh.stock_amount = 5
+                prd_wh.receipt_amount = 28
+                prd_wh.save(update_fields=['stock_amount', 'receipt_amount'])
+                if ProductWareHouseSerial.objects.filter(vendor_serial_number='hp-dk').count() == 0:
+                    new_sn = [
+                        ProductWareHouseSerial(
+                            tenant=company.tenant,
+                            company=company,
+                            vendor_serial_number='hp-dk',
+                            serial_number='hp-dk-01',
+                            product_warehouse=prd_wh
+                        ),
+                        ProductWareHouseSerial(
+                            tenant=company.tenant,
+                            company=company,
+                            vendor_serial_number='hp-dk',
+                            serial_number='hp-dk-02',
+                            product_warehouse=prd_wh
+                        )
+                    ]
+                    ProductWareHouseSerial.objects.bulk_create(new_sn)
+                    prd_wh.product.stock_amount = 5
+                    prd_wh.product.available_amount = -24
+                    prd_wh.product.save(update_fields=['stock_amount', 'available_amount'])
         # đầu kỳ Transceiver
-        ReportInventoryProductWarehouse.objects.create(
-            tenant=company.tenant,
-            company=company,
-            employee_created_id='e37c69ca-c5a0-45ff-9c55-dc0bc37b476e',
-            employee_inherit_id='e37c69ca-c5a0-45ff-9c55-dc0bc37b476e',
-            product_id='1ee9ee35-25af-48f8-95f7-c51fcd4c615c',
-            warehouse_id='bbac9cfc-df1b-4ed4-97c9-a57ac5c94f89',
-            period_mapped_id='5c7423ae29824f338dc5fd2c41b694bf',
-            sub_period_order=3,
-            sub_period_id='5e1c3cccb4c8439d9b3936a69b72b42a',
-            opening_balance_quantity=float(4),
-            opening_balance_value=float(32000000),
-            opening_balance_cost=float(8000000),
-            ending_balance_quantity=float(4),
-            ending_balance_value=float(32000000),
-            ending_balance_cost=float(8000000),
-            for_balance=True
-        )
-        ProductWareHouse.objects.filter(id='a0401285-454e-484b-a54a-5795284f19cd').update(stock_amount=13)
-        Product.objects.filter(id='1ee9ee35-25af-48f8-95f7-c51fcd4c615c').update(stock_amount=16, available_amount=1)
-        lot = ProductWareHouseLot.objects.filter(id='d3d5116b56fa4171a78b71ed53b1cce6').first()
-        if lot:
-            lot.quantity_import = 0
-            lot.save(update_fields=['quantity_import'])
-        lot = ProductWareHouseLot.objects.filter(id='b1d4a1cf-b72b-41d0-817d-0514d49bd352').first()
-        if lot:
-            lot.quantity_import = 13
-            lot.save(update_fields=['quantity_import'])
+        if has_lot:
+            ReportInventoryProductWarehouse.objects.create(
+                tenant=company.tenant,
+                company=company,
+                employee_created_id='e37c69ca-c5a0-45ff-9c55-dc0bc37b476e',
+                employee_inherit_id='e37c69ca-c5a0-45ff-9c55-dc0bc37b476e',
+                product_id='1ee9ee35-25af-48f8-95f7-c51fcd4c615c',
+                warehouse_id='bbac9cfc-df1b-4ed4-97c9-a57ac5c94f89',
+                lot_mapped_id="b1d4a1cfb72b41d0817d0514d49bd352",
+                period_mapped_id='5c7423ae29824f338dc5fd2c41b694bf',
+                sub_period_order=3,
+                sub_period_id='5e1c3cccb4c8439d9b3936a69b72b42a',
+                opening_balance_quantity=float(3),
+                opening_balance_value=float(24000000),
+                opening_balance_cost=float(8000000),
+                ending_balance_quantity=float(3),
+                ending_balance_value=float(24000000),
+                ending_balance_cost=float(8000000),
+                for_balance=True
+            )
+            ReportInventoryProductWarehouse.objects.create(
+                tenant=company.tenant,
+                company=company,
+                employee_created_id='e37c69ca-c5a0-45ff-9c55-dc0bc37b476e',
+                employee_inherit_id='e37c69ca-c5a0-45ff-9c55-dc0bc37b476e',
+                product_id='1ee9ee35-25af-48f8-95f7-c51fcd4c615c',
+                warehouse_id='bbac9cfc-df1b-4ed4-97c9-a57ac5c94f89',
+                lot_mapped_id="d3d5116b56fa4171a78b71ed53b1cce6",
+                period_mapped_id='5c7423ae29824f338dc5fd2c41b694bf',
+                sub_period_order=3,
+                sub_period_id='5e1c3cccb4c8439d9b3936a69b72b42a',
+                opening_balance_quantity=float(1),
+                opening_balance_value=float(8000000),
+                opening_balance_cost=float(8000000),
+                ending_balance_quantity=float(1),
+                ending_balance_value=float(8000000),
+                ending_balance_cost=float(8000000),
+                for_balance=True
+            )
+        else:
+            ReportInventoryProductWarehouse.objects.create(
+                tenant=company.tenant,
+                company=company,
+                employee_created_id='e37c69ca-c5a0-45ff-9c55-dc0bc37b476e',
+                employee_inherit_id='e37c69ca-c5a0-45ff-9c55-dc0bc37b476e',
+                product_id='1ee9ee35-25af-48f8-95f7-c51fcd4c615c',
+                warehouse_id='bbac9cfc-df1b-4ed4-97c9-a57ac5c94f89',
+                period_mapped_id='5c7423ae29824f338dc5fd2c41b694bf',
+                sub_period_order=3,
+                sub_period_id='5e1c3cccb4c8439d9b3936a69b72b42a',
+                opening_balance_quantity=float(4),
+                opening_balance_value=float(32000000),
+                opening_balance_cost=float(8000000),
+                ending_balance_quantity=float(4),
+                ending_balance_value=float(32000000),
+                ending_balance_cost=float(8000000),
+                for_balance=True
+            )
+        if run_update_warehouse:
+            ProductWareHouse.objects.filter(id='a0401285-454e-484b-a54a-5795284f19cd').update(stock_amount=13)
+            Product.objects.filter(id='1ee9ee35-25af-48f8-95f7-c51fcd4c615c').update(stock_amount=16, available_amount=1)
+            lot = ProductWareHouseLot.objects.filter(id='d3d5116b56fa4171a78b71ed53b1cce6').first()
+            if lot:
+                lot.quantity_import = 0
+                lot.save(update_fields=['quantity_import'])
+            lot = ProductWareHouseLot.objects.filter(id='b1d4a1cf-b72b-41d0-817d-0514d49bd352').first()
+            if lot:
+                lot.quantity_import = 13
+                lot.save(update_fields=['quantity_import'])
 
-        ProductWareHouse.objects.filter(id='78fdd83e-6cea-415c-9879-d7dbd6d4a224').update(stock_amount=3)
-        lot = ProductWareHouseLot.objects.filter(id='449f5026-b9dc-4cb0-ade8-4cda3b32018b').first()
-        if lot:
-            lot.quantity_import = 3
-            lot.save(update_fields=['quantity_import'])
+            ProductWareHouse.objects.filter(id='78fdd83e-6cea-415c-9879-d7dbd6d4a224').update(stock_amount=3)
+            lot = ProductWareHouseLot.objects.filter(id='449f5026-b9dc-4cb0-ade8-4cda3b32018b').first()
+            if lot:
+                lot.quantity_import = 3
+                lot.save(update_fields=['quantity_import'])
 
     all_delivery = OrderDeliverySub.objects.filter(
         company_id=company_id, state=2, date_done__year=2024, date_done__month__gte=start_month
@@ -1844,7 +1890,10 @@ def report_rerun(company_id, start_month, run_fix_data=False):
 
     all_doc_sorted = sorted(all_doc, key=lambda x: datetime.fromisoformat(x['date_approved']))
     for doc in all_doc_sorted:
-        print(f"\n----- Run: {doc['id']} | {doc['date_approved']} | {doc['code']} | {doc['type']}")
+        print(f"--- Run id: {doc['id']}")
+        print(f"\tSystem date: {doc['date_approved'].split(' ')[0]}")
+        print(f"\tSystem code: ** {doc['code']} **")
+        print(f"\tType: {doc['type']}")
 
         if doc['type'] == 'delivery':
             instance = OrderDeliverySub.objects.get(id=doc['id'])
@@ -1866,7 +1915,7 @@ def report_rerun(company_id, start_month, run_fix_data=False):
             instance = GoodsTransfer.objects.get(id=doc['id'])
             instance.prepare_data_for_logging(instance)
 
-    print('Done')
+    print('Complete!')
 
 
 def update_product_warehouse_uom_base():
